Made the ban command ban people not in the server.

+++ cogs/mod.py
@@ -15,7 +15,7 @@ class mod(commands.Cog, name="Moderation"):
     @commands.guild_only()
     @commands.has_permissions(ban_members=True)
     @commands.bot_has_permissions(ban_members=True, manage_messages=True)
-    async def ban(self, ctx, member: discord.Member, *, reason="No reason provided"):
+    async def ban(self, ctx, ban_user: discord.User, *, reason="No reason provided"):
         """
         Ban the specified user with the specified reason
         *Reason defaults to __no reason specified__*
@@ -24,10 +24,10 @@ class mod(commands.Cog, name="Moderation"):
         crossmark = '<a:cross:813798012626141185>'
         
         try:
-            if member == ctx.message.author:
+            if ban_user == ctx.message.author:
                 return await ctx.send("You can not ban yourself, please try someone else.")
 
-            if member == self.bot.user:
+            if ban_user == self.bot.user:
                 
                 def check(r, u):
                     return u.id == ctx.author.id and r.message.id == checkmsg.id
@@ -62,17 +62,23 @@ class mod(commands.Cog, name="Moderation"):
                     return await checkmsg.edit(content="Command timed out, canceling...")
        
             botmember = ctx.guild.me
-            if member.top_role > botmember.top_role:
-                return await ctx.send("My role is too low in the hierarchy. Please move it above the highest role the user you are trying to ban has.")
-            await ctx.message.delete()
-            messageok = f"You were banned from `{ctx.guild.name}` with reason:\n\n{reason}"
-            try:
-                await member.send(messageok)
-            except discord.errors.HTTPException:
-                pass
-            await member.ban(reason=f"Moderator: {ctx.message.author} | Reason: {reason}")
-            e = discord.Embed(title=f"{member} was banned | {reason}", color=config.red)
-            await ctx.send(embed=e)
+            if ctx.guild.get_member(ban_user.id) is not None:
+                member = ctx.guild.get_member(ban_user.id)
+                if member.top_role > botmember.top_role:
+                    return await ctx.send("My role is too low in the hierarchy. Please move it above the highest role the user you are trying to ban has.")
+                await ctx.message.delete()
+                messageok = f"You were banned from `{ctx.guild.name}` with reason:\n\n{reason}"
+                try:
+                    await member.send(messageok)
+                except discord.errors.HTTPException:
+                    pass
+                await member.ban(reason=f"Moderator: {ctx.message.author} | Reason: {reason}")
+                e = discord.Embed(title=f"{member} was banned | {reason}", color=config.red)
+                await ctx.send(embed=e)
+            else:
+                await ctx.guild.ban(ban_user, reason=f"Moderator: {ctx.message.author} | Reason: {reason}")
+                e = discord.Embed(title=f"{ban_user} ({ban_user.id}) was banned | {reason}", color=config.red)
+                await ctx.send(embed=e)
         except Exception as e:
             await ctx.send(f"```py\n{e}\n```")
